+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# Master
+
+## Enhancements
+
+- Optional object properties are now always present in generated JSON examples.
+
 # 0.16.1
 
 ## Bug Fixes